added some js-files and organized some code
