+++ js/data/constants.js
@@ -1,14 +1,34 @@
-// URLs................
+// ------------URLs--------------
 export const URL = "https://api.noroff.dev/api/v1/rainy-days";
 
-// -----------Popular product slider
+// -----------Popular product slider--------
 export const productSliderContainer = document.querySelector(".popular_products_slider");
 
-// Product page
-export const imageContainer = document.querySelector(".product-image-container");
+// -----------Product page--------------
 
+export const errorContainer = document.querySelector(".error_container");
+export const imageContainer = document.querySelector(".product-image-container");
 export const productTitleContainer = document.querySelector(".product_choises_headline");
 export const productDescriptionContainer = document.querySelector(".product_choises_list");
 export const productSizeContainer = document.querySelector(".size_option_container");
 export const productPriceContainer = document.querySelector(".price-container");
-export const productProductDetailsContainer = document.querySelector(".product_details");
+export const productDetailsContainer = document.querySelector(".product_details");
+export const wordContainer = document.querySelector(".product_details_list");
+
+export const desiredWords = [
+  "breathable",
+  "sustainable",
+  "durable",
+  "lightweight",
+  "waterproof",
+  "hiking",
+  "stylish",
+  "water-repellent",
+  "outdoor adventure",
+  "outdoor activities",
+  "travel",
+  "packable",
+  "relaxed fit",
+];
+
+// -----------------------------
