+++ product.html
@@ -98,6 +98,7 @@
       </div>
     </header>
     <main class="index_page_content">
+      <div class="error_container"></div>
       <section class="product_top_section">
         <div class="product-image-container">
           <div class="loader"></div>
